リクエスト作成機能をリポジトリメソッドを使用して実装し、エラーハンドリングを追加
